更改了起始价格的变量格式 完善了路线修改引起的产品最低价格变化

+++ ruoyi-system/src/main/java/com/ruoyi/travel/domain/ProductBasic.java
@@ -1,5 +1,6 @@
 package com.ruoyi.travel.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import org.apache.commons.lang3.builder.ToStringBuilder;
@@ -9,7 +10,7 @@ import com.ruoyi.common.core.domain.BaseEntity;
 
 /**
  * 产品信息对象 travel_product_basic
- * 
+ *
  * @author buaa_travel
  * @date 2021-05-12
  */
@@ -56,7 +57,7 @@ public class ProductBasic extends BaseEntity
 
     /** 起始最低价格 */
     @Excel(name = "起始最低价格")
-    private Long priceStart;
+    private BigDecimal priceStart;
 
     /** 订单完成数量 */
     @Excel(name = "订单完成数量")
@@ -78,138 +79,138 @@ public class ProductBasic extends BaseEntity
     @Excel(name = "路线数量")
     private Long routeNumber;
 
-    public void setId(Long id) 
+    public void setId(Long id)
     {
         this.id = id;
     }
 
-    public Long getId() 
+    public Long getId()
     {
         return id;
     }
-    public void setProductName(String productName) 
+    public void setProductName(String productName)
     {
         this.productName = productName;
     }
 
-    public String getProductName() 
+    public String getProductName()
     {
         return productName;
     }
-    public void setProductStatus(Long productStatus) 
+    public void setProductStatus(Long productStatus)
     {
         this.productStatus = productStatus;
     }
 
-    public Long getProductStatus() 
+    public Long getProductStatus()
     {
         return productStatus;
     }
-    public void setProductDay(Long productDay) 
+    public void setProductDay(Long productDay)
     {
         this.productDay = productDay;
     }
 
-    public Long getProductDay() 
+    public Long getProductDay()
     {
         return productDay;
     }
-    public void setEditTime(Date editTime) 
+    public void setEditTime(Date editTime)
     {
         this.editTime = editTime;
     }
 
-    public Date getEditTime() 
+    public Date getEditTime()
     {
         return editTime;
     }
-    public void setProductDesign(String productDesign) 
+    public void setProductDesign(String productDesign)
     {
         this.productDesign = productDesign;
     }
 
-    public String getProductDesign() 
+    public String getProductDesign()
     {
         return productDesign;
     }
-    public void setProductPlan(String productPlan) 
+    public void setProductPlan(String productPlan)
     {
         this.productPlan = productPlan;
     }
 
-    public String getProductPlan() 
+    public String getProductPlan()
     {
         return productPlan;
     }
-    public void setBirthland(String birthland) 
+    public void setBirthland(String birthland)
     {
         this.birthland = birthland;
     }
 
-    public String getBirthland() 
+    public String getBirthland()
     {
         return birthland;
     }
-    public void setDestination(String destination) 
+    public void setDestination(String destination)
     {
         this.destination = destination;
     }
 
-    public String getDestination() 
+    public String getDestination()
     {
         return destination;
     }
-    public void setPriceStart(Long priceStart) 
+    public void setPriceStart(BigDecimal priceStart)
     {
         this.priceStart = priceStart;
     }
 
-    public Long getPriceStart() 
+    public BigDecimal getPriceStart()
     {
         return priceStart;
     }
-    public void setOrderNumber(Long orderNumber) 
+    public void setOrderNumber(Long orderNumber)
     {
         this.orderNumber = orderNumber;
     }
 
-    public Long getOrderNumber() 
+    public Long getOrderNumber()
     {
         return orderNumber;
     }
-    public void setDescription(String description) 
+    public void setDescription(String description)
     {
         this.description = description;
     }
 
-    public String getDescription() 
+    public String getDescription()
     {
         return description;
     }
-    public void setImgUrl(String imgUrl) 
+    public void setImgUrl(String imgUrl)
     {
         this.imgUrl = imgUrl;
     }
 
-    public String getImgUrl() 
+    public String getImgUrl()
     {
         return imgUrl;
     }
-    public void setProductDetailAds(String productDetailAds) 
+    public void setProductDetailAds(String productDetailAds)
     {
         this.productDetailAds = productDetailAds;
     }
 
-    public String getProductDetailAds() 
+    public String getProductDetailAds()
     {
         return productDetailAds;
     }
-    public void setRouteNumber(Long routeNumber) 
+    public void setRouteNumber(Long routeNumber)
     {
         this.routeNumber = routeNumber;
     }
 
-    public Long getRouteNumber() 
+    public Long getRouteNumber()
     {
         return routeNumber;
     }
